backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==23.2
djoser==2.1.0
# asgiref==3.7.2
drf-extra-fields==3.7.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
Pillow==9.0.0
PyYAML==6.0.1
python-dotenv==1.0.0
pycparser==2.21
flake8==6.0.0
# pytz==2023.3
# sqlparse==0.4.4
# typing_extensions==4.6.3